Add ratings from `CreateRating FC`
There is probably a way to patch in just the new rating rather than
calling for all ratings to be requested and rendered again. Perhaps
with local state, or pushing it into all ratings and updating that
one just throgh `useEffect()`, but this is fine for now.

Addresses #26 (Create `CreateRating` component)
